refactor(web): split admin teams route components
